everything.

+++ src/java/growthcraft/grapes/client/renderer/RenderGrape.java
@@ -2,62 +2,56 @@ package growthcraft.grapes.client.renderer;
 
 import growthcraft.core.util.RenderUtils;
 import net.minecraft.block.Block;
-
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.world.IBlockAccess;
-
 import net.minecraftforge.fml.client.registry.RenderingRegistry;
 
-public class RenderGrape implements ISimpleBlockRenderingHandler
-{
-	public static final int id = RenderingRegistry.getNextAvailableRenderId();
+public class RenderGrape implements ISimpleBlockRenderingHandler {
+    public static final int id = RenderingRegistry.getNextAvailableRenderId();
 
-	@Override
-	public void renderInventoryBlock(Block block, int metadata, int modelID, RenderBlocks renderer) {}
+    @Override
+    public void renderInventoryBlock(Block block, int metadata, int modelID, RenderBlocks renderer) {
+    }
 
-	@Override
-	public boolean renderWorldBlock(IBlockAccess world, BlockPos pos, Block block, int modelId, RenderBlocks renderer)
-	{
-		if (modelId == id)
-		{
-			final Tessellator tessellator = Tessellator.instance;
-			tessellator.setBrightness(block.getMixedBrightnessForBlock(world, x, y, z));
-			tessellator.setColorOpaque_F(1.0F, 1.0F, 1.0F);
-			final IIcon icon = block.getIcon(0, 0);
-			final double d = 0.0625D;
+    @Override
+    public boolean renderWorldBlock(IBlockAccess world, BlockPos pos, Block block, int modelId, RenderBlocks renderer) {
+        if (modelId == id) {
+            final Tessellator tessellator = Tessellator.instance;
+            tessellator.setBrightness(block.getMixedBrightnessForBlock(world, x, y, z));
+            tessellator.setColorOpaque_F(1.0F, 1.0F, 1.0F);
+            final IIcon icon = block.getIcon(0, 0);
+            final double d = 0.0625D;
 
-			long random = (long)(x * 3129871) ^ (long)y * 116129781L ^ (long)z;
-			random = random * random * 42317861L + random * 11L;
+            long random = x * 3129871 ^ (long) y * 116129781L ^ z;
+            random = random * random * 42317861L + random * 11L;
 
-			double minX = (double)x + 3*d;
-			minX += ((double)((float)(random >> 16 & 15L) / 15.0F) - 0.5D) * 0.5D;
-			final double maxX = minX + 10*d;
-			double minY = (double)y + 9*d;
-			minY += ((double)((float)(random >> 20 & 15L) / 15.0F) - 1.0D) * 0.2D;
-			final double maxY = minY + 10*d;
-			double minZ = (double)z + 3*d;
-			minZ += ((double)((float)(random >> 24 & 15L) / 15.0F) - 0.5D) * 0.5D;
-			final double maxZ = minZ + 10*d;
+            double minX = (double) x + 3 * d;
+            minX += ((double) ((float) (random >> 16 & 15L) / 15.0F) - 0.5D) * 0.5D;
+            final double maxX = minX + 10 * d;
+            double minY = (double) y + 9 * d;
+            minY += ((double) ((float) (random >> 20 & 15L) / 15.0F) - 1.0D) * 0.2D;
+            final double maxY = minY + 10 * d;
+            double minZ = (double) z + 3 * d;
+            minZ += ((double) ((float) (random >> 24 & 15L) / 15.0F) - 0.5D) * 0.5D;
+            final double maxZ = minZ + 10 * d;
 
-			final double minU = (double)icon.getMinU();
-			final double maxU = (double)icon.getMaxU();
-			final double minV = (double)icon.getMinV();
-			final double maxV = (double)icon.getMaxV();
+            final double minU = (double) icon.getMinU();
+            final double maxU = (double) icon.getMaxU();
+            final double minV = (double) icon.getMinV();
+            final double maxV = (double) icon.getMaxV();
 
-			RenderUtils.drawCrossSquaresAlongY(tessellator, minX, maxX, minY, maxY, minZ, maxZ, minU, maxU, minV, maxV);
-		}
-		return true;
-	}
+            RenderUtils.drawCrossSquaresAlongY(tessellator, minX, maxX, minY, maxY, minZ, maxZ, minU, maxU, minV, maxV);
+        }
+        return true;
+    }
 
-	@Override
-	public boolean shouldRender3DInInventory(int modelID)
-	{
-		return false;
-	}
+    @Override
+    public boolean shouldRender3DInInventory(int modelID) {
+        return false;
+    }
 
-	@Override
-	public int getRenderId()
-	{
-		return id;
-	}
+    @Override
+    public int getRenderId() {
+        return id;
+    }
 }
